requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1